Multiplayer works?
I think i have what  i need to make multiplayer work but i dont have
another controller to test it with. But if it does work then it is done.
Car class and race manager will be to be changed a little once car
select is in but for now it is a good place holder