Fixed turing bug and made it so the car flips when overturned
I changed the value of the dead zone for the analog sticks and that
fixed the turning bug, also I made it so that the car resets to it's
ORIGINAL rotation when overturned, I tried to make it so it still faces
the same direction but that created errors. I will work on this further
to try and fix it.